modele, szkielet security - wszystko troche rozgrzebane

+++ backend/src/main/resources/db/migration/V1.0.4__create_transaction_table.sql
@@ -4,7 +4,7 @@ CREATE TABLE transaction
     address_id INTEGER NOT NULL,
     source_address VARCHAR(35) NOT NULL,
     destination_address VARCHAR(35) NOT NULL,
-    amount BIGINT NOT NULL,
+    amount DECIMAL NOT NULL,
     created_at TIMESTAMP DEFAULT now() NOT NULL,
     status SMALLINT NOT NULL,
     confirmations INTEGER DEFAULT 0 NOT NULL,
